refactor(client server): reorganize includes and improve code readability across multiple files

+++ client/src/game.cpp
@@ -1,19 +1,20 @@
 #include "Game.hpp"
 
-#include <SFML/Graphics.hpp>
 #include "inputKey.hpp"
+
+#include <SFML/Graphics.hpp>
+
+#include <chrono>
+#include <iostream>
 #include <memory>
 #include <string>
-#include <iostream>
-#include <chrono>
 
 Game::Game(ThreadSafeQueue<GameToNetwork::Message>& game_to_net,
            ThreadSafeQueue<NetworkToGame::Message>& net_to_game)
-    : game_to_network_queue_(game_to_net)
-    , network_to_game_queue_(net_to_game)
-    , window_(sf::VideoMode(WINDOW_WIDTH, WINDOW_HEIGHT), "R-TYPE Client")
-    , is_running_(true)
-{
+    : game_to_network_queue_(game_to_net),
+      network_to_game_queue_(net_to_game),
+      window_(sf::VideoMode(WINDOW_WIDTH, WINDOW_HEIGHT), "R-TYPE Client"),
+      is_running_(true) {
     std::cout << "[Game] Starting of the R-TYPE client..." << std::endl;
 
     window_.setFramerateLimit(FRAMERATE);
@@ -64,8 +65,7 @@ void Game::handle_input() {
 
     if (input_mask != 0) {
         game_to_network_queue_.push(
-            GameToNetwork::Message(GameToNetwork::MessageType::SendInput, input_mask)
-        );
+            GameToNetwork::Message(GameToNetwork::MessageType::SendInput, input_mask));
     }
 }
 
@@ -100,10 +100,18 @@ void Game::process_network_messages() {
 
                     incoming.shape.setSize(sf::Vector2f(30.f, 30.f));
                     switch (incoming.type) {
-                        case 0x01: incoming.shape.setFillColor(sf::Color::Green); break;
-                        case 0x02: incoming.shape.setFillColor(sf::Color::Red); break;
-                        case 0x03: incoming.shape.setFillColor(sf::Color::Yellow); break;
-                        default: incoming.shape.setFillColor(sf::Color::White); break;
+                        case 0x01:
+                            incoming.shape.setFillColor(sf::Color::Green);
+                            break;
+                        case 0x02:
+                            incoming.shape.setFillColor(sf::Color::Red);
+                            break;
+                        case 0x03:
+                            incoming.shape.setFillColor(sf::Color::Yellow);
+                            break;
+                        default:
+                            incoming.shape.setFillColor(sf::Color::White);
+                            break;
                     }
 
                     next[id] = std::move(incoming);
@@ -138,24 +146,32 @@ void Game::render() {
         auto curr_t = e.curr_time;
 
         if (curr_t > prev_t) {
-            const double total_ms = std::chrono::duration_cast<std::chrono::duration<double, std::milli>>(curr_t - prev_t).count();
-            const double elapsed_ms = std::chrono::duration_cast<std::chrono::duration<double, std::milli>>(render_time - prev_t).count();
-            double alpha = (total_ms > 0.0) ? (elapsed_ms / total_ms) : 1.0;
-            if (alpha <= 0.0) {
+            const float total_ms =
+                std::chrono::duration_cast<std::chrono::duration<float, std::milli>>(curr_t -
+                                                                                     prev_t)
+                    .count();
+            const float elapsed_ms =
+                std::chrono::duration_cast<std::chrono::duration<float, std::milli>>(render_time -
+                                                                                     prev_t)
+                    .count();
+            float alpha = (total_ms > 0.0f) ? (elapsed_ms / total_ms) : 1.0f;
+            if (alpha <= 0.0f) {
                 draw_x = e.prev_x;
                 draw_y = e.prev_y;
-            } else if (alpha >= 1.0) {
-                const double after_ms = std::chrono::duration_cast<std::chrono::duration<double>>(render_time - curr_t).count();
-                if (after_ms > 0.0) {
-                    draw_x = e.x + static_cast<float>(e.vx * after_ms);
-                    draw_y = e.y + static_cast<float>(e.vy * after_ms);
+            } else if (alpha >= 1.0f) {
+                const float after_ms =
+                    std::chrono::duration_cast<std::chrono::duration<float>>(render_time - curr_t)
+                        .count();
+                if (after_ms > 0.0f) {
+                    draw_x = e.x + (e.vx * after_ms);
+                    draw_y = e.y + (e.vy * after_ms);
                 } else {
                     draw_x = e.x;
                     draw_y = e.y;
                 }
             } else {
-                draw_x = static_cast<float>(e.prev_x + (e.x - e.prev_x) * alpha);
-                draw_y = static_cast<float>(e.prev_y + (e.y - e.prev_y) * alpha);
+                draw_x = e.prev_x + (e.x - e.prev_x) * alpha;
+                draw_y = e.prev_y + (e.y - e.prev_y) * alpha;
             }
         } else {
             draw_x = e.x;
